# Improve proxy performance and connection handling