--- a/proxy/proxy_test.go
+++ b/proxy/proxy_test.go
@@ -190,3 +190,41 @@ func TestBackendDialError(t *testing.T) {
 	p.Shutdown()
 	p.Wait()
 }
+
+func TestCopyData(t *testing.T) {
+	size := 16 /* bytes */
+	proxy := New(nil, 60*time.Second, nil, &testLogger{}, LogEverything, false)
+
+	srcIn, srcOut := net.Pipe()
+	dstIn, dstOut := net.Pipe()
+	defer func() {
+		srcIn.Close()
+		srcOut.Close()
+		dstIn.Close()
+		dstOut.Close()
+	}()
+
+	go func() {
+		proxy.copyData(dstIn, srcOut)
+	}()
+
+	input := make([]byte, size)
+	for i := 0; i < size; i++ {
+		input[i] = byte(i)
+	}
+
+	_, _ = srcIn.Write(input)
+	srcIn.Close()
+
+	output := make([]byte, size)
+	n, err := dstOut.Read(output)
+	if n != size {
+		t.Fatalf("expected %d bytes, got %d instead", size, n)
+	}
+	if err != nil && err != io.EOF && !isClosedConnectionError(err) {
+		t.Fatalf("got unexpected error: %v", err)
+	}
+	if !bytes.Equal(input, output) {
+		t.Fatalf("input and output were different after copy")
+	}
+}