--- a/tests/test-client-handles-client-closes-connection.py
+++ b/tests/test-client-handles-client-closes-connection.py
@@ -32,6 +32,14 @@ if __name__ == "__main__":
         pair.validate_closing_client_closes_server(
             "1: client closed -> server closed")
 
+        pair = SocketPair(TcpClient(13001), TlsServer('server', 'root', 13002))
+        pair.validate_can_send_from_client(
+            "hello world", "2: client -> server")
+        pair.validate_can_send_from_server(
+            "hello world", "2: server -> client")
+        pair.validate_half_closing_client_closes_server(
+            "2: client closed -> server closed")
+
         print_ok("OK")
     finally:
         terminate(ghostunnel)