Improvements to the core proxy package:
* Set up a pool of reusable buffers to reduce allocations
* Close only corresponding read, write side of connection when copying data
* Set deadline (via configurable timeout) for max conn lifetime after other side closes
* New flag for setting max conn lifetime regardless of circumstances (defaults to zero)
* Add benchmarks for the proxy package to test allocs

Before this change:
```
goos: darwin
goarch: arm64
pkg: github.com/ghostunnel/ghostunnel/proxy
cpu: Apple M1
BenchmarkCopyData/1_bytes-8         	  155918	      6713 ns/op	   36730 B/op	      31 allocs/op
BenchmarkCopyData/2_bytes-8         	  158008	      6809 ns/op	   36731 B/op	      31 allocs/op
BenchmarkCopyData/4_bytes-8         	  175996	      6766 ns/op	   36733 B/op	      31 allocs/op
BenchmarkCopyData/8_bytes-8         	  148226	      7429 ns/op	   36737 B/op	      31 allocs/op
BenchmarkCopyData/16_bytes-8        	  151615	      6977 ns/op	   36745 B/op	      31 allocs/op
BenchmarkCopyData/32_bytes-8        	  156847	      7013 ns/op	   36761 B/op	      31 allocs/op
BenchmarkCopyData/64_bytes-8        	  156166	      7003 ns/op	   36793 B/op	      31 allocs/op
BenchmarkCopyData/128_bytes-8       	  154147	      6953 ns/op	   36857 B/op	      31 allocs/op
BenchmarkCopyData/256_bytes-8       	  154548	      7029 ns/op	   36985 B/op	      31 allocs/op
BenchmarkCopyData/512_bytes-8       	  149385	      8058 ns/op	   37241 B/op	      31 allocs/op
BenchmarkCopyData/1024_bytes-8      	  145123	      7664 ns/op	   37753 B/op	      31 allocs/op
BenchmarkCopyData/2048_bytes-8      	  126390	      8548 ns/op	   38777 B/op	      31 allocs/op
BenchmarkCopyData/4096_bytes-8      	  104260	     10800 ns/op	   40825 B/op	      31 allocs/op
BenchmarkCopyData/8192_bytes-8      	   74418	     14866 ns/op	   44921 B/op	      31 allocs/op
BenchmarkCopyData/16384_bytes-8     	   51024	     22794 ns/op	   53114 B/op	      31 allocs/op
BenchmarkCopyData/32768_bytes-8     	   28970	     41048 ns/op	   69499 B/op	      31 allocs/op
PASS
ok  	github.com/ghostunnel/ghostunnel/proxy	21.014s
```

After this change:
```
goos: darwin
goarch: arm64
pkg: github.com/ghostunnel/ghostunnel/proxy
cpu: Apple M1
BenchmarkCopyData/1_bytes-8         	  279038	      4150 ns/op	    3962 B/op	      30 allocs/op
BenchmarkCopyData/2_bytes-8         	  296374	      4082 ns/op	    3962 B/op	      30 allocs/op
BenchmarkCopyData/4_bytes-8         	  300432	      4058 ns/op	    3964 B/op	      30 allocs/op
BenchmarkCopyData/8_bytes-8         	  289056	      4045 ns/op	    3968 B/op	      30 allocs/op
BenchmarkCopyData/16_bytes-8        	  293391	      4102 ns/op	    3977 B/op	      30 allocs/op
BenchmarkCopyData/32_bytes-8        	  291940	      4160 ns/op	    3992 B/op	      30 allocs/op
BenchmarkCopyData/64_bytes-8        	  293884	      4108 ns/op	    4025 B/op	      30 allocs/op
BenchmarkCopyData/128_bytes-8       	  290247	      4248 ns/op	    4088 B/op	      30 allocs/op
BenchmarkCopyData/256_bytes-8       	  286138	      4253 ns/op	    4216 B/op	      30 allocs/op
BenchmarkCopyData/512_bytes-8       	  274270	      4412 ns/op	    4473 B/op	      30 allocs/op
BenchmarkCopyData/1024_bytes-8      	  258865	      4791 ns/op	    4985 B/op	      30 allocs/op
BenchmarkCopyData/2048_bytes-8      	  196558	      6031 ns/op	    6010 B/op	      30 allocs/op
BenchmarkCopyData/4096_bytes-8      	  146282	      8132 ns/op	    8059 B/op	      30 allocs/op
BenchmarkCopyData/8192_bytes-8      	   99974	     12115 ns/op	   12155 B/op	      30 allocs/op
BenchmarkCopyData/16384_bytes-8     	   59104	     19751 ns/op	   20350 B/op	      30 allocs/op
BenchmarkCopyData/32768_bytes-8     	   31572	     37970 ns/op	   36747 B/op	      30 allocs/op
PASS
ok  	github.com/ghostunnel/ghostunnel/proxy	21.051s
```